fix(checker): a local export specifier on an import-alias cycle reports its own TS2303
When a local `export { X as Y }` (no `from` clause) re-exports an import alias
that is part of a circular alias chain, tsc reports TS2303 at the specifier as
well as at the import; tsz now does too, through a companion scan in the
checker's circular-alias walk.

`bind_export_declaration` resolves a local specifier's name in the current scope
and publishes the exported name onto that *existing* symbol, so a cyclic alias
and every local re-export of it share one `SymbolId` and the walk sees a shorter
cycle than tsc's. tsc gives each specifier its own alias symbol on the same
cycle. The extra anchors are therefore recovered syntactically from the
declaring container, exactly as the `export =` companion scan above it already
does and for the same binder reason.

Only the specifier the cycle actually re-enters under is a member of it: for
`export { X as B, X as C }` tsc resolves `C` through an already-resolved alias
and never revisits it. The walk now carries the export name of the hop that
closed the cycle so the scan can tell a member from a branch.

Also splits the TS2303 detection out of `module_checker.rs` into
`module_checker/circular_alias.rs`; the file was 1949 lines and the change
pushed it past the 2000-line architecture cap.
